ctr: move fetch,fetch-object,push-object to content
