+++ cmd/ctr/main.go
@@ -78,14 +78,11 @@ containerd CLI
 		containers.Command,
 		content.Command,
 		events.Command,
-		fetchCommand,
-		fetchObjectCommand,
 		images.Command,
 		namespacesCmd.Command,
 		pprofCommand,
 		pullCommand,
 		pushCommand,
-		pushObjectCommand,
 		rootfsCommand,
 		runCommand,
 		snapshot.Command,
